@@ -22,15 +22,15 @@
     "build-config": "node build-config.js"
   },
   "dependencies": {
-    "@angular/animations": "^13.3.0",
-    "@angular/common": "^13.3.0",
-    "@angular/compiler": "^13.3.0",
-    "@angular/core": "^13.3.0",
-    "@angular/forms": "^13.3.0",
-    "@angular/platform-browser": "^13.3.0",
-    "@angular/platform-browser-dynamic": "^13.3.0",
-    "@angular/router": "^13.3.0",
-    "@ionic/angular": "^6.0.13",
+    "@angular/animations": "^14.1.0",
+    "@angular/common": "^14.1.0",
+    "@angular/compiler": "^14.1.0",
+    "@angular/core": "^14.1.0",
+    "@angular/forms": "^14.1.0",
+    "@angular/platform-browser": "^14.1.0",
+    "@angular/platform-browser-dynamic": "^14.1.0",
+    "@angular/router": "^14.1.0",
+    "@ionic/angular": "^6.1.15",
     "@ionic/storage-angular": "^3.0.6",
     "@materia-ui/ngx-monaco-editor": "^6.0.0",
     "@start9labs/argon2": "^0.1.0",
@@ -47,10 +47,10 @@
     "marked": "^4.0.0",
     "monaco-editor": "^0.33.0",
     "mustache": "^4.2.0",
-    "ng-qrcode": "^6.0.0",
+    "ng-qrcode": "^7.0.0",
     "patch-db-client": "file: ../../../patch-db/client",
     "pbkdf2": "^3.1.2",
-    "rxjs": "^6.6.7",
+    "rxjs": "^7.5.6",
     "swiper": "^8.2.4",
     "ts-matches": "^5.1.7",
     "tslib": "^2.3.0",
@@ -58,10 +58,10 @@
     "zone.js": "^0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.1.4",
-    "@angular/cli": "^13.1.4",
-    "@angular/compiler-cli": "^13.3.0",
-    "@angular/language-service": "^13.3.0",
+    "@angular-devkit/build-angular": "^14.1.0",
+    "@angular/cli": "^14.1.0",
+    "@angular/compiler-cli": "^14.1.0",
+    "@angular/language-service": "^14.1.0",
     "@ionic/cli": "^6.19.0",
     "@types/aes-js": "^3.1.1",
     "@types/dompurify": "^2.3.3",
@@ -74,7 +74,7 @@
     "@types/uuid": "^8.3.1",
     "husky": "^4.3.8",
     "lint-staged": "^12.3.7",
-    "ng-packagr": "^13.3.0",
+    "ng-packagr": "^14.1.0",
     "node-html-parser": "^5.3.3",
     "prettier": "^2.6.1",
     "raw-loader": "^4.0.2",
