@@ -1,9 +1,10 @@
 import { first } from 'rxjs/operators'
 import { PatchDbService } from 'src/app/services/patch-db/patch-db.service'
 import { UIMarketplaceData } from 'src/app/services/patch-db/data-model'
+import { firstValueFrom } from 'rxjs'
 
 export function getMarketplace(
   patch: PatchDbService,
 ): Promise<UIMarketplaceData> {
-  return patch.watch$('ui', 'marketplace').pipe(first()).toPromise()
+  return firstValueFrom(patch.watch$('ui', 'marketplace'))
 }
